requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework_simplejwt
PyJWT
python-dotenv
pytz
sqlparse
tzdata
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
python-dateutil
icalendar
requests
celery[redis]
redis
